requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
flake8==6.0.0
Flask==2.3.2
flask-cors==6.0.1
fpdf==1.7.2
google-ai-generativelanguage==0.6.2
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.10.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.0.1
pyparsing==3.2.3
pytest==7.4.0
python-dotenv==1.0.0
python-pptx==0.6.21
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.5
